requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
python-dotenv==1.2.2
qdrant-client==1.17.1
ollama==0.6.1
httpx==0.28.1
weasyprint==68.1